requirements.txt
openai
psycopg2-binary
sqlalchemy
pgvector
pandas
python-dotenv
tqdm
streamlit
faiss-cpu
numpy